requirements.txt
openai==1.54.4
langchain==0.3.7
langchain-openai==0.2.8
chromadb==0.5.20
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic==2.10.2
requests==2.32.3
pytest==8.3.3
streamlit==1.39.0
plotly==5.24.1
fastapi>=0.95.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
langchain-community
tiktoken
xlsxwriter
fpdf
matplotlib